+++ src/routes/favorites.js
