+++ src/routes/index.js
@@ -1,25 +1,25 @@
 const { Router } = require("express");
 
-const router_admin = require("./router_admin");
-const router_cities = require("./router_cities");
-const router_favorites = require("./router_favorites");
-const router_feedback = require("./router_feedback");
-const router_interested = require("./router_interested");
-const router_mercado = require("./router_mercado");
-const router_membership = require("./router_memberships");
-const router_properties = require("./router_properties");
+const admin = require("./admin");
+const cities = require("./cities");
+const favorites = require("./favorites");
+const feedback = require("./feedback");
+const interested = require("./interested");
+const mercado = require("./mercado");
+const membership = require("./memberships");
+const properties = require("./properties");
 const users = require("./users");
 
 const router = Router();
 
-router.use("/", router_admin);
-router.use("/", router_cities);
-router.use("/", router_favorites);
-router.use("/", router_feedback);
-router.use("/", router_interested);
-router.use("/", router_mercado);
-router.use("/", router_membership);
-router.use("/", router_properties);
+router.use("/", admin);
+router.use("/", cities);
+router.use("/", favorites);
+router.use("/", feedback);
+router.use("/", interested);
+router.use("/", mercado);
+router.use("/", membership);
+router.use("/", properties);
 router.use("/user", users);
 
 module.exports = router;
